feat: simplify project
BREAKING CHANGE: update and simplify API.
Many of the added helper functions have bloated the project leaving
unneeded complexity in which function to use. This update should
encourage the use of the standard libary while using this libary only
for the core of interacting with objects through the reader and writer
it provides along with some simple bucket operations such as List, Exist
and Delete.